--- app/components/AppLayout.tsx
@@ -1,15 +1,0 @@
-import React from "react";
-import Footer from "~/components/Footer";
-import Header from "~/components/Header";
-
-export default function AppLayout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <div className="flex min-h-screen flex-col">
-      <Header />
-      {children}
-      <Footer />
-    </div>
-  );
-}
