+++ react-router.config.ts
@@ -4,4 +4,5 @@ export default {
   // Config options...
   // Server-side render by default, to enable SPA mode set this to `false`
   ssr: true,
+  appDirectory: "./src/app",
 } satisfies Config;
