+++ src/app/routes/home.tsx
@@ -1,0 +1,13 @@
+import type { Route } from "./+types/home";
+import HomePage from "~/pages/home";
+
+export function meta({}: Route.MetaArgs) {
+  return [
+    { title: "New React Router App" },
+    { name: "description", content: "Welcome to React Router!" },
+  ];
+}
+
+export default function HomeRoute() {
+  return <HomePage />;
+}
