+++ src/widgets/Footer.tsx
@@ -1,9 +1,11 @@
 import React from "react";
 
-export default function Footer() {
+function Footer() {
   return (
     <footer className="flex items-center justify-center bg-slate-900 p-4">
       <p className="text-background">Made by Danylkovych Dmytro</p>
     </footer>
   );
 }
+
+export { Footer };
