+++ src/widgets/index.ts
@@ -1,0 +1,8 @@
+import { AppLayout } from "./AppLayout";
+import { CategoryCard } from "./CategoryCard";
+import { CategoryGrid } from "./CategoryGrid";
+import { Footer } from "./Footer";
+import { Header } from "./Header";
+import { PopularQuiz } from "./PopularQuiz";
+
+export { AppLayout, CategoryCard, CategoryGrid, Footer, Header, PopularQuiz };
